01_group_A/requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0
cryptography>=3.4.8

